Cleaned up, simplified
Moved config files to top level -- we tend to only have a Dockerfile and
a config file, let's make it easy to see how simple this is.

Tried to add a cluster.name to elasticsearch but this caused logstash to
whine about not joining the cluster: it worked OK but warnings were
confusing.